# openapi/openapi.spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/task": {
      "post": {
        "operationId": "TaskController_create",
        "summary": "Create new Task",
        "description": "This endpoint creates a new task and links it to existing categories and tags",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Task information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The task was succesfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      },
      "get": {
        "operationId": "TaskController_findAll",
        "summary": "Get all Tasks",
        "description": "This endpoint returns a list of all tasks with the name of the related tags, categories and users",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/task/{id}": {
      "get": {
        "operationId": "TaskController_findOne",
        "summary": "Get one Task with related entities in detail",
        "description": "This endpoint returns a single tag with all the related entities in detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      },
      "patch": {
        "operationId": "TaskController_update",
        "summary": "Update one Task",
        "description": "This endpoint updates all the task info and its relationships",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Updated Task information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      },
      "delete": {
        "operationId": "TaskController_remove",
        "summary": "Remove one Task",
        "description": "This endpoint removes a task",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/category": {
      "post": {
        "operationId": "CategoryController_create",
        "summary": "Create new Category",
        "description": "This endpoint creates a new category",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Category name and description",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The category was succesfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "CategoryController_findAll",
        "summary": "Get all Categories",
        "description": "This endpoint returns a list of all categories",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/category/{id}": {
      "get": {
        "operationId": "CategoryController_findOne",
        "summary": "Get one Category",
        "description": "This endpoint returns a single category",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "patch": {
        "operationId": "CategoryController_update",
        "summary": "Update one Category",
        "description": "This endpoint updates the category name and description",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Updated Category information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "operationId": "CategoryController_remove",
        "summary": "Remove one category",
        "description": "This endpoint removes a category (with onCascade: \"SET NULL\" for task)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/tag": {
      "post": {
        "operationId": "TagController_create",
        "summary": "Create new Tag",
        "description": "This endpoint creates a new tag",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Tag name and color",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The tag was succesfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tag"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "TagController_findAll",
        "summary": "Get all Tags",
        "description": "This endpoint returns a list of all tags",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/tag/{id}": {
      "get": {
        "operationId": "TagController_findOne",
        "summary": "Get one Tag",
        "description": "This endpoint returns a single tag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "patch": {
        "operationId": "TagController_update",
        "summary": "Update one Tag",
        "description": "This endpoint updates the tag name and color",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Updated tag information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "operationId": "TagController_remove",
        "summary": "Remove one Tag",
        "description": "This endpoint removes a tag (with onCascade: \"SET NULL\" for task)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Login",
        "description": "This endpoint receives an email and a password and after validating them, returns a jwt token with the encrypted user infos",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "User email and password",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User succesfully logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Register",
        "description": "This endpoint receives an email and a password and creates a new user if email is not already present",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "User email and password",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User succesfully created and logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    }
  },
  "info": {
    "title": "Todo API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateTaskDto": {
        "type": "object",
        "properties": {}
      },
      "Task": {
        "type": "object",
        "properties": {}
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {}
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {}
      },
      "Category": {
        "type": "object",
        "properties": {}
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {}
      },
      "CreateTagDto": {
        "type": "object",
        "properties": {}
      },
      "Tag": {
        "type": "object",
        "properties": {}
      },
      "UpdateTagDto": {
        "type": "object",
        "properties": {}
      },
      "LoginRequestDTO": {
        "type": "object",
        "properties": {}
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {}
      },
      "RegisterRequestDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}